Spherical harmonic basis generator works. Added pre-multiplied domega.
git-svn-id: https://s.snth.net/svn/rlk/thumb@1633 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ src/ogl-sh-basis.hpp
@@ -27,10 +27,10 @@ namespace ogl
 
         GLuint object;
 
-        void fill(float *, const float *,
-                           const float *,
-                           const float *,
-                           const float *) const;
+        void fill(float *, const double *,
+                           const double *,
+                           const double *,
+                           const double *) const;
 
     public:
